Both instanced and modelcache rendering
